update: implemented some further OOP. not working still

+++ DungeonCrafter/core/src/dev/teamcyan/dungeoncrafter/classes/Pos.java
@@ -1,31 +1,31 @@
 package dev.teamcyan.dungeoncrafter.classes;
 
 public class Pos {
-    private float x;
-    private float y;
+    private int x;
+    private int y;
 
-    public Pos (float x, float y)
+    public Pos (int x, int  y)
     {
         this.x=x;
         this.y=y;
     }
 
-    public void setX(float x)
+    public void setX(int x)
     {
         this.x=x;
     }
 
-    public void setY(float y)
+    public void setY(int  y)
     {
         this.y=y;
     }
 
-    public float getY()
+    public int  getY()
     {
        return this.y; 
     }
 
-    public float getX()
+    public int getX()
     {
        return this.x; 
     }
